Create README.md
In this particular project, I worked on the inaugural corpora from the nltk in 
Python. I looked at the following speeches of the Presidents of the United States of 
America: 
1. President Franklin D. Roosevelt in 1941 
2. President John F. Kennedy in 1961 
3. President Richard Nixon in 1973  
The dataset consists of speeches of three US Presidents. The objective of the project is to 
perform text analysis on the data to find the most frequent words and they mean and at last 
we will be analyzing the data and creating word clouds. The learning outcome in this project 
is Test Analysis, text cleaning by removing Stop words, special characters and creating word 
cloud.